+++ backend/src/main/java/me/fairygel/fbook/service/GenreService.java
@@ -1,5 +1,6 @@
 package me.fairygel.fbook.service;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
 import me.fairygel.fbook.dto.genre.GenreDTO;
 import me.fairygel.fbook.dto.genre.GenreIndexViewDTO;
@@ -17,37 +18,41 @@ public class GenreService {
     private final GenreCrudRepository genreCrudRepository;
     private final GenreMapper mapper;
 
+    private static final String NO_GENRE = "No genre with id = ";
+
     public void create(GenreDTO genreDTO) {
         Genre genre = mapper.genreDtoToGenre(genreDTO);
         genreCrudRepository.save(genre);
     }
 
-    // we will try to find genre with id, then, if it doesn't exist, we will try to get empty genre
-    // if he also isn't exists, throw an error
     public GenreDTO read(Long id) {
-        Genre genre = genreCrudRepository
-                .findById(id).orElse(
-                        genreCrudRepository.findById(0L)
-                                .orElseThrow(IllegalStateException::new)
-                );
+        if (id == 0L) throw new EntityNotFoundException(NO_GENRE + id);
+
+        Genre genre = genreCrudRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(NO_GENRE + id));
         return mapper.genreToGenreDto(genre);
     }
 
     public GenreDTO update(Long id, GenreDTO genreDTO) {
+        if (id == 0L) throw new EntityNotFoundException(NO_GENRE + id);
+
         Genre genre = mapper.genreDtoToGenre(genreDTO);
 
-        Genre updatedGenre = genreCrudRepository.updateById(id, genre).orElseThrow(IllegalAccessError::new);
+        Genre updatedGenre = genreCrudRepository.updateById(id, genre)
+                .orElseThrow(() -> new EntityNotFoundException(NO_GENRE + id));
 
         return mapper.genreToGenreDto(updatedGenre);
     }
 
     public void delete(Long id) {
+        if (id == 0L) throw new EntityNotFoundException(NO_GENRE + id);
+
         genreCrudRepository.deleteById(id);
     }
 
     public Set<GenreIndexViewDTO> index() {
         Set<Genre> genres = new HashSet<>();
-        genreCrudRepository.findAll().forEach(genres::add);
+        genreCrudRepository.findAll().forEach(g -> {if (g.getId() != 0L) genres.add(g);});
 
         return mapper.genresToIndex(genres);
     }
